Emergency fix to get Merge Bot functional again
If someone force pushes while a merge run is in progress, the merge bot
falls over. This fix will should change the error we're seeing into a
no-op, and we should silently skip over that merge run.

In the future there should at least be a comment posted to GitHub
explaining what just happened.